Add files via upload
So, Car Price Prediction is machine learning model which is built on Random Forest Algorithm, it is able to predict car's market value, which can help both buyers and sellers. So In this Project, we are going to predict the Price of Used Cars using various attributes such as kilometers driven, age, number of previous owners etc.